requirements.txt
agno
google-search-results
python-dotenv
groq
googlesearch-python
pycountry
google
google-genai
google-generativeai
duckduckgo-search